+++ lisp/cider-mode.el
@@ -441,6 +441,8 @@ If invoked with a prefix ARG eval the expression after inserting it."
      ["Inspect" cider-inspect]
      ["Toggle var tracing" cider-toggle-trace-var]
      ["Toggle ns tracing" cider-toggle-trace-ns]
+     ["List traced" cider-list-traced]
+     ["Untrace all" cider-untrace-all]
      "--"
      ["Debug top-level form" cider-debug-defun-at-point
       :keys "\\[universal-argument] \\[cider-eval-defun-at-point]"]